Implemented remote repo and added test for get methods

+++ app/src/main/java/com/alligator/weatherforecast/data/model/response/response_model/Coordinate.kt
@@ -6,7 +6,7 @@ import kotlinx.serialization.Serializable
 @Serializable
 data class Coordinate(
     @SerialName("lon")
-    val longitude:Float,
+    val longitude:Double,
     @SerialName("lat")
-    val latitude:Float
+    val latitude:Double
 )
